feat(workers): consolidate waveformLod, colorScopes, and subtitleParser into mediaAnalysis.worker
Merges three separate worker isolates into a single persistent media analysis
domain worker, eliminating ~80 MB of idle V8 heap overhead.

mediaAnalysis.worker.ts handles:
  - Multi-LOD audio waveform pyramid build and viewport slice (zero-copy Float32Array)
  - Real-time 60fps color scopes via OffscreenCanvas (histogram, vectorscope, parade)
  - Subtitle/transcript tokenisation (SRT, VTT, ASS, Whisper word alignment)

Removes: waveformLod.worker.ts, colorScopes.worker.ts, subtitleParser.worker.ts
